Split helper functions into seperate files
