+++ src/lib/classNames.ts
@@ -1,0 +1,3 @@
+const classNames = (...classes: string[]) => classes.filter(Boolean).join(" ");
+
+export default classNames;
